% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_Function.R
\name{Posterior_Function}
\alias{Posterior_Function}
\title{Posterior Function}
\usage{
Posterior_Function(Prior_Weight, prior, Sampling_Weight)
}
\arguments{
\item{input}{4) A vector of frequencies for sampling data}
}
\description{
This function plots a posterior curve with Base R
}
\examples{
Posterior_Function
}
\keyword{Bayesian}
\keyword{Posterior,}
