% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior_Function.R
\name{Prior_Function}
\alias{Prior_Function}
\title{Prior Function}
\usage{
Prior_Function(prior_data)
}
\arguments{
\item{input}{the vector of data to be plotted; mean and standard deviation can be varied within the function}
}
\description{
This function plots a normal curve of data
}
\keyword{curve}
\keyword{density}
\keyword{prior,}
