% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingFunction.R
\name{SamplingFunction}
\alias{SamplingFunction}
\title{Sampling Function}
\usage{
SamplingFunction(SampleData, NumbTrials)
}
\arguments{
\item{input}{the number of trials}
}
\description{
This function plots a probability density function of binary data options
}
\keyword{density}
\keyword{probability,}
